backend/requirements.txt
fastapi
uvicorn[standard]
motor
passlib
python-jose[cryptography]
python-dotenv
pydantic-settings
email-validator
python-multipart
